python_tests/pyproject.toml
[project]
name = "routiium-tests"
version = "0.1.0"
description = "Python integration tests for routiium"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.uv]
dev-dependencies = []
